test: rename test_strategies
