+++ tests/test_strategies.py
